api/requirements.txt
# =============================================================================
# NEWS INTELLIGENCE SYSTEM - REQUIREMENTS
# =============================================================================
# 
# Core dependencies for the News Intelligence System
# Install with: pip install -r requirements.txt
#
# =============================================================================

# =============================================================================
# CORE DEPENDENCIES (Required for all deployments)
# =============================================================================
# Web framework
Flask==2.3.3
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
Werkzeug==2.3.7

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# HTTP and networking
requests==2.31.0
urllib3==2.0.7

# RSS and content processing
feedparser==6.0.10
beautifulsoup4==4.12.2
lxml==4.9.3

# Data processing
pandas==2.1.3
numpy==1.25.2

# Utilities
python-dateutil==2.8.2
pytz==2023.3
click==8.1.7

# =============================================================================
# MONITORING DEPENDENCIES (Optional - for monitoring profile)
# =============================================================================
# System monitoring
psutil==5.9.6

# Metrics and monitoring
prometheus-client==0.17.1

# =============================================================================
# SECURITY DEPENDENCIES (Required for production)
# =============================================================================
# Authentication and security
PyJWT==2.8.0
cryptography==41.0.7
bcrypt==4.1.2

# =============================================================================
# ENVIRONMENT-SPECIFIC DEPENDENCIES
# =============================================================================
# Local development
python-dotenv==1.0.0

# Production
gunicorn==21.2.0
uvicorn==0.24.0

# =============================================================================
# VERSION CONSTRAINTS AND COMPATIBILITY
# =============================================================================
# Ensure compatibility with Python 3.11+
# All packages are tested with Python 3.11.7
